Address IN EXPLAIN doc and JIT int-fold minor items
Two low-risk minor items from the PR #7320 level-3 review.

InLongFunctionFactory.plan() javadoc claimed the merged mixed-width
value list was "byte-for-byte identical to the original single-set
plan". That is inaccurate: the mixed-width path sorts the merged list,
so its output is in ascending order and does not match the hash order
of the single-set branches. The comment now states this and warns that
an EXPLAIN assertion over a mixed-width IN list must expect ascending
order. Doc-only; no behavior change.

tryFoldConstantArithI4() parsed a leaf constant with
(int) Numbers.parseLong(token), which silently truncates a LONG-range
literal. A leaf here is always in INT range because any out-of-INT leaf
routes to the I8/widen branch, so the truncation was only latent.
Numbers.parseInt() throws NumericException on an out-of-range token
instead, so a future violation of that invariant makes the caller
cleanly fall back to descending the subtree as per-op IR rather than
emitting a silently wrong immediate.

+++ core/src/main/java/io/questdb/jit/CompiledFilterIRSerializer.java
@@ -2562,7 +2562,12 @@ public class CompiledFilterIRSerializer implements PostOrderTreeTraversalAlgo.Vi
             throw NumericException.INSTANCE;
         }
         if (node.type == ExpressionNode.CONSTANT) {
-            return (int) Numbers.parseLong(node.token);
+            // A leaf constant here is always in INT range: an out-of-INT leaf routes
+            // to the I8/widen branch instead of the I4 fold. parseInt throws on an
+            // out-of-range token rather than silently truncating a LONG-range literal,
+            // so if that invariant is ever violated the caller cleanly falls back to
+            // descending the subtree as per-op IR.
+            return Numbers.parseInt(node.token);
         }
         if (node.type != ExpressionNode.OPERATION) {
             throw NumericException.INSTANCE;
